Unified Map integration and unit test runner
