+++ integration/map_integration_test.sh
@@ -4,7 +4,6 @@ INTEGRATION_DIR="$( cd "$( dirname "$0" )" && pwd )"
 . "${INTEGRATION_DIR}"/functions.sh
 
 echo "Building code"
-go build ${GOFLAGS} github.com/google/trillian/cmd/createtree/
 go build ${GOFLAGS} github.com/google/trillian/server/trillian_map_server/
 
 yes | "${TRILLIAN_PATH}/scripts/resetdb.sh"
@@ -17,20 +16,10 @@ RPC_SERVER_PID=$!
 wait_for_server_startup ${RPC_PORT}
 TO_KILL+=(${RPC_SERVER_PID})
 
-echo "Provision map"
-TEST_TREE_ID=$(./createtree \
-  --admin_server="localhost:${RPC_PORT}" \
-  --tree_type=MAP \
-  --pem_key_path=${TRILLIAN_PATH}/testdata/map-rpc-server.privkey.pem \
-  --pem_key_password=towel \
-  --hash_strategy=TEST_MAP_HASHER \
-  --signature_algorithm=ECDSA)
-echo "Created tree ${TEST_TREE_ID}"
-
 echo "Running test"
 cd "${INTEGRATION_DIR}"
 set +e
-go test -run ".*LiveMap.*" --timeout=5m ./ --map_id ${TEST_TREE_ID} --map_rpc_server="localhost:${RPC_PORT}" --pubkey=${TRILLIAN_PATH}/testdata/map-rpc-server.pubkey.pem
+go test --timeout=5m ./maptest  --map_rpc_server="localhost:${RPC_PORT}" 
 RESULT=$?
 set -e
